requirements.txt
streamlit
pandas
scikit-learn
fpdf
pdfminer.six
docx2txt
